SFER-89: moved mud audio to other script
