+++ SferoTestVR/Assets/Code/Scripts/Obstacles/MudAudioCotroller.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6f45882e4942c6a408c9ff4da23ae878
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
