+++ SferoTestVR/Assets/Code/Scripts/Player/Player.cs
@@ -17,7 +17,7 @@ public class Player : MonoBehaviour
     
 
     [SerializeField] protected AudioSource playerRollingAudioSource;
-    [SerializeField] protected AudioSource playerMudAudioSource;
+    
     [SerializeField] protected AudioSource hitAudioSource;
 	[SerializeField] protected float minSpeedSound = 0.5f;
     [SerializeField] protected float maxSpeedSound = 5.0f;
@@ -43,7 +43,6 @@ public class Player : MonoBehaviour
     private float startTime;
 
     private bool isOnFloor = false;
-    private bool isOnMud = false;
 
     /// <summary> 
     /// Bool whether it's the player, or the game controlling the sphere 
@@ -97,9 +96,6 @@ public class Player : MonoBehaviour
         }
 
         CalculateRollingVolume();
-        CalculateMudVolume();
-
-
 	}
 
     virtual public void SetPlayerControlsSelf(bool playerControlsSelf)
@@ -261,11 +257,6 @@ public class Player : MonoBehaviour
 		{
 			isOnFloor = true;
 		}
-
-		if (collision.gameObject.CompareTag("Mud") && !isOnMud)
-		{
-			isOnMud = true;
-		}
 	}
 
 	private void OnCollisionExit(Collision collision)
@@ -273,11 +264,6 @@ public class Player : MonoBehaviour
 		if (collision.gameObject.CompareTag("Floor"))
 		{
             isOnFloor = false;
-		}
-
-		if (collision.gameObject.CompareTag("Mud"))
-		{
-			isOnMud = false;
 		}
 	}
 
@@ -402,31 +388,6 @@ public class Player : MonoBehaviour
 		else
 		{
 			playerRollingAudioSource.Stop();
-		}
-	}
-
-
-	/// <summary>
-	/// Calculates volume and pitch of mud sound
-	/// </summary> 
-	protected void CalculateMudVolume()
-	{
-		float currentPlayerSpeed = rigidbody.velocity.magnitude;
-        float normalizedSpeed = Mathf.Clamp01((currentPlayerSpeed - minSpeedSound/2.0f) / (maxSpeedSound - minSpeedSound/2.0f));
-
-		playerMudAudioSource.volume = normalizedSpeed;
-        playerMudAudioSource.pitch = Mathf.Lerp(1.0f, 2.0f, normalizedSpeed);
-
-		if ( isOnMud)
-		{
-			if (!playerMudAudioSource.isPlaying)
-			{
-				playerMudAudioSource.Play();
-			}
-		}
-		else
-		{
-			playerMudAudioSource.Stop();
 		}
 	}
